افزودن حذف افزودن انتخاب همه به آپشن های لیست اصلاح کامیت قبلی

+++ ParsKyanCrm.Application/Dtos/Users/DataFormQuestionsDto.cs
@@ -22,6 +22,6 @@ namespace ParsKyanCrm.Application.Dtos.Users
         public string QuestionName { get; set; }
         public string QuestionType { get; set; }       
         public int? QuestionOrder { get; set; }
-        public float? Score { get; set; }
+        public int? Score { get; set; }
     }
 }
